Lens

Theory

The solution is

Approximations

Thin lens approximation

Constructor

Jolab.Lens — Method

Lens(T, f, na, ref)

Initializes a multilayer structure.

T number type specifing data precision. Ex: Float32, BigFloat, ... The default is Float64;

f - defines the lens focal length in meters. A function or interpolation object can be used to describe a focal length wavelength dependent.

na - defines the lens numerical aperture. A function or interpolation object can be used to describe a focal length wavelength dependent.

ref - ReferenceFrame specifying the lens position and orientation. The focal planes are located at the plus and minus the focal length.

Examples:

lens = Lens(10E-3, 0.5, ReferenceFrame(0,0,0,0,0))

f(λ) = 10E-3 * λ
na(λ) = .1 * λ
lens = Lens(f, na, ReferenceFrame(0,0,0,0,0))

Implementation

Missing docstring.

Missing docstring for lightinteraction(::Lens, ::FieldAngularSpectrumScalar). Check Documenter's build log for details.

Jolab.coefficient_general — Method

(field_back, field_forward)= coefficient_general(lens, angspe)

Calculates the field after propagating through the lens assuming the incident field direction of propagation. The model propagates from a focal plane to the other focal plane. The order is based on the light direction of propagation.

Type: Transmission matrices is diagonal. No reflection matrix is computed (the model does not assume light reflection by the lens)

Time: very short; scales with length(angspe.nsx_X) length(angspe.nsy_Y)

RAM: very small; scales with length(angspe.nsx_X) length(angspe.nsy_Y)

Convergence: sampling of angspe.nsx_X and angspe.nsy_Y

Applications

Example of systems studied with similar models:

Fabry-Pérot etalon

Aberations induced by a coverslip

Reference

If you use this model, please cite: Marques, Dylan M., et al. "Modelling fabry-pérot etalons illuminated by focussed beams." Optics express 28.5 (2020): 7691-7706.

Additional reading:

Goodman, Joseph W. Introduction to Fourier optics. Roberts and Company Publishers, 2005.

Wolf, Emil. "Electromagnetic diffraction in optical systems-I. An integral representation of the image field." Proceedings of the Royal Society of London. Series A. Mathematical and Physical Sciences 253.1274 (1959): 349-357.

Richards, Bernard, and Emil Wolf. "Electromagnetic diffraction in optical systems, II. Structure of the image field in an aplanatic system." Proceedings of the Royal Society of London. Series A. Mathematical and Physical Sciences 253.1274 (1959): 358-379.

Novotny, Lukas, and Bert Hecht. Principles of nano-optics. Cambridge university press, 2012.