Multilayer Structure

A Multilayer structure or stratified media is a stack of parallel films.

Theory

The solution is based on calculating the reflected and transmitted plane waves at each interface and the phase change due to light propagation inside each layer. The solution is a geometric sum which considers all the internal round trips generated inside the multilayer structure as illustrated in the figure above.

Approximations

None. The implementation solves Maxwell equations for any type of multilayer structure and any incident beam.

Constructor

Jolab.MultilayerStructure — Method

MultilayerStructure(T, n, h, ref)

Initializes a multilayer structure.

T number type specifing data precision. Ex: Float32, BigFloat, ... The default is Float64;

n - vector specifying the refractive index of each layer. A function or interpolation object can be used to describe a refractive index wavelength dependent.

h - vector specifying the thickness of each layer. The layer with refractive index n[i] has a thickness of h[i-1] meters;

ref - ReferenceFrame specifying the multilayer position and orientation. The first interface of the multilayer structure is intersect the reference frame.

Examples:

mls = MultilayerStructure([1, 2, 3], [100E-9], ReferenceFrame(0,0,0.,0,0))

n_1(λ) = √(complex(8.393 + .14383 / ((λ*1E6)^2 - 0.2421^2) + 4430.99 / ((λ*1E6)^2 - 36.71^2)))
mls = MultilayerStructure([1, n_1], [], ReferenceFrame(0,0,0.,0,0))

Implementation

Jolab.lightinteraction — Method

lightinteraction(::MultilayerStructure, ::FieldAngularSpectrumScalar)

Calculates the reflected and transmitted fields from a multilayer structure for an incident angular spectrum

Time: very short; scales with Nx Ny

RAM: None

Convergence sampling of nsx and nsy

Jolab.coefficient_general — Method

 coefficient_general(::MultilayerStructure, ::FieldAngularSpectrumScalar)

Calculates the scattering matrix of a multilayer structure for an incident angular spectrum

Type: Transmission and reflection matrices are diagonal

Time: very short; scales with Nx Ny

RAM: very small; scales with Nx Ny

Convergence sampling of nsx and nsy

Applications

Example of systems studied with similar models:

Fabry-Pérot etalon

Aberations induced by a coverslip

Reference

If you use this model, please cite: Marques, Dylan M., et al. "Modelling fabry-pérot etalons illuminated by focussed beams." Optics express 28.5 (2020): 7691-7706.

Additional reading:

Chapter 2 of Fujiwara, Hiroyuki. Spectroscopic ellipsometry: principles and applications. John Wiley & Sons, 2007.

Van de Nes, A. S., et al. "Calculation of the vectorial field distribution in a stratified focal region of a high numerical aperture imaging system." Optics Express 12.7 (2004): 1281-1293.